# validator/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Validator Service",
    "version": "0.2.0",
    "description": "Real-time game-action validation service (BYOSnap). Game-facing match RPCs (InitMatch, ValidateAction, CompleteMatch) are a gRPC service (moveborne.validator.v1.ValidatorService, container port 8080) defined in validator/protos/, reached in production through the Snapser Hermes WebSocket (MESSAGE_TYPE_SNAP_API_PROXY) and locally through the validator's own Hermes-emulation endpoint (GET /hermes/ws?token=<player-id>, WebSocket upgrade). The HTTP surface below carries the platform probe, status, state-history debug tooling, and the MCP debugging endpoint.\n\nAuthentication: each operation declares `x-snapser-auth-types` (user|api-key|internal). Operations that also set `x-snapser-auth-passthrough: true` are public \u2014 the gateway does not require a session token. The validator binds user-context requests to the gateway-validated User-Id header (must match player_id / the match owner); the same binding applies to gRPC metadata."
  },
  "servers": [
    {
      "url": "https://gateway.snapser.com/c4n1awfs/v1/byosnap-validator",
      "description": "Snapser dev snapend gateway (c4n1awfs)"
    },
    {
      "url": "http://localhost:5555",
      "description": "Local development"
    }
  ],
  "tags": [
    {
      "name": "Service",
      "description": "Service info and health"
    },
    {
      "name": "Match",
      "description": "Match lifecycle and state history"
    },
    {
      "name": "Realtime",
      "description": "Hermes-envelope WebSocket (local emulation of the gateway's Hermes endpoint)"
    },
    {
      "name": "Debug",
      "description": "MCP inspection/debugging interface"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Service info",
        "description": "Returns a human-readable summary of the service and its primary endpoints.",
        "operationId": "getServiceInfo",
        "responses": {
          "200": {
            "description": "Service information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceInfo"
                }
              }
            }
          }
        },
        "x-snapser-auth-types": [
          "user",
          "api-key",
          "internal"
        ],
        "x-snapser-auth-passthrough": true
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Health check",
        "description": "Liveness/readiness probe. Always returns 200 with a timestamp when the process is up. Used as the BYOSnap readiness probe path.",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        },
        "x-snapser-auth-types": [
          "user",
          "api-key",
          "internal"
        ],
        "x-snapser-auth-passthrough": true
      }
    },
    "/api/status": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Runtime status",
        "description": "Reports server name, version, process uptime, and current Socket.IO connection count.",
        "operationId": "getStatus",
        "responses": {
          "200": {
            "description": "Runtime status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        },
        "x-snapser-auth-types": [
          "user",
          "api-key",
          "internal"
        ],
        "x-snapser-auth-passthrough": true
      }
    },
    "/api/match/init-from-history": {
      "post": {
        "tags": [
          "Match"
        ],
        "summary": "Initialize a match from a saved state history",
        "description": "Creates a replay match from a previously saved state history. Provide either `history_file_id` (looked up in the history store or fixtures) or inline `history_data`. Optionally start from a specific `start_from_index`.",
        "operationId": "initMatchFromHistory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidatorInitFromHistoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Match registered from history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "match_id": {
                      "type": "string"
                    },
                    "state_count": {
                      "type": "integer"
                    },
                    "current_move_index": {
                      "type": "integer"
                    },
                    "expires_at": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error (missing player_id, no history source, empty/invalid history, or no state at start index)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "History not found in store or filesystem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-snapser-auth-types": [
          "user",
          "api-key",
          "internal"
        ],
        "x-snapser-auth-passthrough": true
      }
    },
    "/api/match/save-history": {
      "post": {
        "tags": [
          "Match"
        ],
        "summary": "Save a state history",
        "description": "Persists an array of state snapshots to the history store and returns a generated id that can later be replayed via /api/match/init-from-history.",
        "operationId": "saveHistory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveHistoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "History saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or empty states array",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-snapser-auth-types": [
          "api-key",
          "internal"
        ]
      }
    },
    "/api/match/load-history/{id}": {
      "get": {
        "tags": [
          "Match"
        ],
        "summary": "Load a saved state history",
        "description": "Returns the stored state snapshots for a previously saved history id.",
        "operationId": "loadHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "History id returned by /api/match/save-history",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "History id is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "History not found or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-snapser-auth-types": [
          "api-key",
          "internal"
        ]
      }
    },
    "/mcp": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "MCP JSON-RPC endpoint",
        "description": "Model Context Protocol (StreamableHTTP) endpoint for inspecting and debugging match state. Exposes tools: list_matches, get_match_state, get_state_history, simulate_action, clear_match. Requires `Accept: application/json, text/event-stream`. NOT public \u2014 keep behind authentication, as it can mutate match state.",
        "operationId": "mcpJsonRpc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonRpcRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response (delivered as an SSE event stream)",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-snapser-auth-types": [
          "api-key",
          "internal"
        ]
      }
    },
    "/hermes/ws": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "summary": "Hermes-envelope WebSocket (local dev emulation)",
        "description": "WebSocket upgrade carrying binary protobuf hermes.ClientMessage/ServerMessage frames \u2014 the same envelope the Snapser Hermes endpoint speaks. snap_api_request payloads are dispatched to the ValidatorService RPC handlers. The ?token= query param is the self-stamped player id locally (gateway headers win when present). Not meaningfully reachable through the gateway (use the real Hermes endpoint there); only partially describable in OpenAPI.",
        "operationId": "hermesWs",
        "x-snapser-auth-types": [
          "user",
          "api-key",
          "internal"
        ],
        "x-snapser-auth-passthrough": true,
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Local self-stamped player id (the deployed Hermes endpoint takes the session token here)."
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols (WebSocket upgrade)"
          },
          "401": {
            "description": "Missing token / gateway identity"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Standard error envelope returned by HTTP endpoints.",
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message"
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
      "ServiceInfo": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Moveborne Validator Service"
          },
          "endpoints": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "health": "/health",
              "status": "/api/status",
              "matchInit": "/api/match/init",
              "socketio": "/socket.io/"
            }
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Epoch milliseconds"
          }
        },
        "required": [
          "status",
          "timestamp"
        ]
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "server": {
            "type": "string",
            "example": "validator"
          },
          "version": {
            "type": "string",
            "example": "0.1.0"
          },
          "uptime": {
            "type": "number",
            "format": "double",
            "description": "Process uptime in seconds"
          },
          "connections": {
            "type": "integer",
            "description": "Active Socket.IO client count"
          }
        },
        "required": [
          "server",
          "version",
          "uptime",
          "connections"
        ]
      },
      "Tile": {
        "type": "object",
        "description": "A single board cell. Shape is defined by the game logic package; extra fields are allowed.",
        "additionalProperties": true,
        "properties": {
          "isEmpty": {
            "type": "boolean"
          },
          "value": {
            "type": "integer"
          }
        }
      },
      "Board": {
        "type": "object",
        "description": "Game board. Extra fields from the game logic package are allowed.",
        "additionalProperties": true,
        "properties": {
          "tiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tile"
            }
          }
        }
      },
      "SynchronizedGameState": {
        "type": "object",
        "description": "Deterministic, hashable game state shared between client and validator. Defined by @spyre-io/moveborne-logic; only the fields the validator reads are described here and additional fields are allowed.",
        "additionalProperties": true,
        "properties": {
          "moveIndex": {
            "type": "integer",
            "description": "Monotonic action index"
          },
          "score": {
            "type": "integer"
          },
          "gameStatus": {
            "type": "string",
            "description": "e.g. IN_PROGRESS, GAME_OVER",
            "example": "IN_PROGRESS"
          },
          "board": {
            "$ref": "#/components/schemas/Board"
          },
          "rngIndices": {
            "type": "object",
            "description": "Per-stream RNG consumption indices.",
            "additionalProperties": true
          },
          "randomSeeds": {
            "description": "Seeds for the deterministic RNG streams.",
            "additionalProperties": true
          }
        },
        "required": [
          "moveIndex",
          "board",
          "gameStatus",
          "rngIndices"
        ]
      },
      "StateHistorySnapshot": {
        "type": "object",
        "properties": {
          "moveIndex": {
            "type": "integer"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Epoch milliseconds"
          },
          "state": {
            "$ref": "#/components/schemas/SynchronizedGameState"
          }
        },
        "required": [
          "moveIndex",
          "state"
        ]
      },
      "ValidatorInitFromHistoryRequest": {
        "type": "object",
        "description": "Provide exactly one of history_file_id or history_data.",
        "properties": {
          "history_file_id": {
            "type": "string",
            "description": "Id of a stored/fixture history"
          },
          "history_data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StateHistorySnapshot"
            }
          },
          "start_from_index": {
            "type": "integer",
            "description": "moveIndex to start from; defaults to the last state"
          },
          "player_id": {
            "type": "string"
          }
        },
        "required": [
          "player_id"
        ]
      },
      "SaveHistoryRequest": {
        "type": "object",
        "properties": {
          "states": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/StateHistorySnapshot"
            }
          }
        },
        "required": [
          "states"
        ]
      },
      "SaveHistoryResponse": {
        "type": "object",
        "properties": {
          "history_id": {
            "type": "string"
          },
          "state_count": {
            "type": "integer"
          }
        },
        "required": [
          "history_id",
          "state_count"
        ]
      },
      "LoadHistoryResponse": {
        "type": "object",
        "properties": {
          "history_id": {
            "type": "string"
          },
          "state_count": {
            "type": "integer"
          },
          "states": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StateHistorySnapshot"
            }
          }
        },
        "required": [
          "history_id",
          "state_count",
          "states"
        ]
      },
      "JsonRpcRequest": {
        "type": "object",
        "description": "JSON-RPC 2.0 request envelope used by the MCP endpoint.",
        "properties": {
          "jsonrpc": {
            "type": "string",
            "enum": [
              "2.0"
            ]
          },
          "id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ]
          },
          "method": {
            "type": "string",
            "example": "tools/list"
          },
          "params": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "jsonrpc",
          "method"
        ]
      }
    }
  }
}